Refactor ComponentInspectorView
